+++ src/main/radclient.c
@@ -55,7 +55,7 @@ static bool done = true;
 static bool print_filename = false;
 
 static fr_ipaddr_t client_ipaddr;
-static int client_port = 0;
+static uint16_t client_port = 0;
 
 static int sockfd;
 static int last_used_id = -1;
